Replace native Select with Combobox
