+++ src/main.ts
@@ -31,8 +31,8 @@ export { default as FormFieldUrlAliasInput } from './components/FormFieldUrlAlia
 export { default as NavigationGrid } from './components/NavigationGrid';
 export { default as LoadableImage } from './components/LoadableImage';
 export { default as DropdownMenu } from './components/DropdownMenu/index.vue';
-export { default as BaseSelect } from './components/BaseSelect/index.vue';
-export { default as FormFieldSelect } from './components/FormFieldSelect/index.vue';
+export { default as BaseSelect } from './components/BaseSelect';
+export { default as FormFieldSelect } from './components/FormFieldSelect';
 export { default as FormFieldSingleSelect } from './components/FormFieldSingleSelect/index.vue';
 export { default as TabList } from './components/TabList';
 export { default as ButtonField } from './components/ButtonField';
