+++ src/plugin.ts
@@ -28,8 +28,8 @@ import { ToastProvider } from './components/Toast';
 import NavigationGrid from './components/NavigationGrid';
 import LoadableImage from './components/LoadableImage';
 import DropdownMenu from './components/DropdownMenu/index.vue';
-import BaseSelect from './components/BaseSelect/index.vue';
-import FormFieldSelect from './components/FormFieldSelect/index.vue';
+import BaseSelect from './components/BaseSelect';
+import FormFieldSelect from './components/FormFieldSelect';
 import FormFieldSingleSelect from './components/FormFieldSingleSelect/index.vue';
 import TabList from './components/TabList';
 import ColorInput from './components/ColorInput';
